Chore: 이메일 인증 시스템 변수명, 변수 위치 수정 (#68)
