+++ src/main/java/org/sesac/slopedbe/member/controller/MemberController.java
@@ -13,15 +13,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import lombok.AllArgsConstructor;
+
 @RequestMapping("/api/users")
 @RestController
+@AllArgsConstructor
 public class MemberController {
 
     private final MemberService memberService;
-
-    public MemberController(MemberService memberService) {
-        this.memberService = memberService;
-    }
 
     @GetMapping("/duplicate-check")
     public ResponseEntity<Boolean> checkDuplicateEmail(@RequestParam String email) {
